+++ composer.json
@@ -17,8 +17,7 @@
         "doctrine/common": "~2.3",
         "doctrine/inflector": "~1.0",
         "leedavis81/drest-common": "dev-master",
-        "zendframework/zend-code": "~2.0",
-        "symfony/yaml": "2.*"
+        "zendframework/zend-code": "~2.0"
     },
     "require-dev": {
         "phpunit/phpunit": "~4.1",
@@ -27,7 +26,8 @@
         "symfony/http-foundation": "~2.0"
     },
     "suggest": {
-        "symfony/http-foundation": "If you're not using Drest with an existing framework adapters (see documentation) then you MUST include this" 
+        "symfony/http-foundation": "If you're not using Drest with an existing framework adapters (see documentation) then you MUST include this",
+        "symfony/yaml": "If you plan on using the YAML driver you MUST include this."
     },    
     "bin": ["bin", "bin/drest.php"],
     "autoload": {
